# Extend Str:: helpers after/before and add new Str:: helpers afterLast/beforeLast