--- a/src/Illuminate/Database/Console/Factories/stubs/factory.stub
+++ b/src/Illuminate/Database/Console/Factories/stubs/factory.stub
@@ -2,8 +2,8 @@
 
 /* @var $factory \Illuminate\Database\Eloquent\Factory */
 
-use Faker\Generator as Faker;
 use NamespacedDummyModel;
+use Faker\Generator as Faker;
 
 $factory->define(DummyModel::class, function (Faker $faker) {
     return [