--- a/src/Illuminate/Support/Collection.php
+++ b/src/Illuminate/Support/Collection.php
@@ -420,12 +420,7 @@ class Collection implements ArrayAccess, Arrayable, Countable, IteratorAggregate
 
         $uniqueItems = $items->unique(null, $strict);
 
-        $compare = $strict ? function ($a, $b) {
-            return $a === $b;
-        }
-        : function ($a, $b) {
-            return $a == $b;
-        };
+        $compare = $this->duplicateComparator($strict);
 
         $duplicates = new static;
 
@@ -449,6 +444,25 @@ class Collection implements ArrayAccess, Arrayable, Countable, IteratorAggregate
     public function duplicatesStrict($callback = null)
     {
         return $this->duplicates($callback, true);
+    }
+
+    /**
+     * Get the comparison function to detect duplicates.
+     *
+     * @param  bool  $strict
+     * @return \Closure
+     */
+    protected function duplicateComparator($strict)
+    {
+        if ($strict) {
+            return function ($a, $b) {
+                return $a === $b;
+            };
+        }
+
+        return function ($a, $b) {
+            return $a == $b;
+        };
     }
 
     /**