--- a/src/Illuminate/Support/Str.php
+++ b/src/Illuminate/Support/Str.php
@@ -122,7 +122,7 @@ class Str
     }
 
     /**
-     * Get the portion of a string between delimiters
+     * Get the portion of a string between delimiters.
      *
      * @param  string  $subject
      * @param  string  $after
@@ -130,14 +130,15 @@ class Str
      * @param  bool    $case_insensitive
      * @return string
      */
-    public static function between($subject, $after, $before = null, $case_insensitive = false) {
+    public static function between($subject, $after, $before = null, $case_insensitive = false)
+    {
         $before = $before === null ? $after : $before;
 
         return self::before(self::after($subject, $after, $case_insensitive), $before, $case_insensitive);
     }
 
     /**
-     * Get the largest portion of a string between delimiters
+     * Get the largest portion of a string between delimiters.
      *
      * @param  string  $subject
      * @param  string  $after
@@ -145,14 +146,15 @@ class Str
      * @param  bool    $case_insensitive
      * @return string
      */
-    public static function betweenGreedy($subject, $after, $before = null, $case_insensitive = false) {
+    public static function betweenGreedy($subject, $after, $before = null, $case_insensitive = false)
+    {
         $before = $before === null ? $after : $before;
 
         return self::beforeLast(self::after($subject, $after, $case_insensitive), $before, $case_insensitive);
     }
 
     /**
-     * Get the last portion of a string between delimiters
+     * Get the last portion of a string between delimiters.
      *
      * @param  string  $subject
      * @param  string  $after
@@ -160,7 +162,8 @@ class Str
      * @param  bool    $case_insensitive
      * @return string
      */
-    public static function betweenLazy($subject, $after, $before = null, $case_insensitive = false) {
+    public static function betweenLazy($subject, $after, $before = null, $case_insensitive = false)
+    {
         $before = $before === null ? $after : $before;
 
         return self::afterLast(self::beforeLast($subject, $before, $case_insensitive), $after, $case_insensitive);