Small but nice change. Non-breaking.

Add case-insensitive support for `Str::after` and `Str::before` allowing you to search before/after an occurrence of a substring, ignoring case, by passing `true` as the third argument.

And add new helpers `Str::afterLast` and `Str::beforeLast` which act similar to `Str::after` and `Str::before`, which find the the first occurrence of the substring, and instead find the last occurrence of the substring. This also has case-insensitive support by passing `true` as the third argument.

---

Ran some benchmarks for a million iterations (variance in results were ± 0.1s:

Context:
```php
$string = 'foo bar foo BAR foo bar foo BAR foo';
$find = 'BAR';
$wontFind = 'Whoops';
```

Current benchmarks:
- `Str::before($string, $find)`: 2.020s
- `Str::after($string, $find)`: 2.831s
- `Str::before($string, $wontFind)`: 1.912s
- `Str::after($string, $wontFind)`: 2.685s

Proposed-change benchmarks _(basically the same code for these)_:
- `Str::before($string, $find)`: 1.973s
- `Str::after($string, $find)`: 2.819s
- `Str::before($string, $wontFind)`: 1.890s
- `Str::after($string, $wontFind)`: 2.591s

New feature benchmarks:
- `Str::beforeLast($string, $find)`: 2.714s
- `Str::afterLast($string, $find)`: 2.786s
- `Str::beforeLast($string, $find, true)`: 2.957s
- `Str::afterLast($string, $find, true)`: 3.061s
- `Str::beforeLast($string, $wontFind)`: 1.930s
- `Str::afterLast($string, $wontFind)`: 1.881s
- `Str::before($string, $find, true)`: 3.916s
- `Str::after($string, $find, true)`: 3.022s

I tried a few different ways of supporting case-sensitive and case-insensitive lookups concurrently, by using `$case_insensitive ? stripos() : strpos()` (and `$case_insensitive ? strripos() : strrpos()`), however the "Current benchmarks" increased quite a bit -- because obviously the current code is optimal. This is why the code is split across a few lines using the existing code if case-sensitive (i.e. `explode`), otherwise `stripos()` if case-insensitive. Not as clean as it could be, but the benchmarks are comparable and is non-breaking.

--

Example Usage:

```php
// after (insensitive)
Str::after('This and this AND this', 'AND', true); // returns: ' this AND this'

// before (insensitive)
Str::after('This and this AND this', 'AND', true); // returns: 'This '

// afterLast
Str::afterLast('\\App\\Http\\Controllers\\UserController', '\\'); // returns: 'UserController'

// beforeLast
Str::beforeLast('\\App\\Http\\Controllers\\UserController', '\\'); // returns: '\\App\\Http\\Controllers'
```

I'm sure you can imagine what the insensitive versions of `afterLast` and `beforeLast` do.