@sisve 2019-04-14
Have you tested this with multibyte strings, like UTF-8 encoded characters?