@devcircus 2019-04-14
This will at least need to target master as the method signature has been changed