@UnDeAdYeTii 2019-04-14
No I hadn't. Have now, all appears working with the respective `mb_` functions instead :)